fix: ensure expected minimal node version for readable error; instead of a non descriptive node error
#30
